day 7 tests passing

+++ 07/include/amp_loop.h
@@ -9,9 +9,9 @@ struct Executioner {
   struct Machine* machines[N_PHASES];
   int phases[N_PHASES];
 };
-int* ingest(int* size);
-int find_max_thrust(int phases[N_PHASES], int l, int r, int list[], int size, int* max);
-struct Executioner* init_executioner(int prog[], int prog_size);
+long* ingest(int* size);
+int find_max_thrust(int phases[N_PHASES], int l, int r, long list[], int size, int* max);
+struct Executioner* init_executioner(long prog[], int prog_size);
 int compute_val(struct Executioner* ex, int phases[N_PHASES]);
 
 #endif
