+++ 07/src/io.c
@@ -7,27 +7,27 @@
 int machine_reader(void* data_ptr, enum OpSig* sig) {
   struct Store* store = (struct Store*)data_ptr;
 
-  int ret = store->send_input ? store->current_val : store->phases[store->amp_n++];
+  long ret = store->send_input ? store->current_val : store->phases[store->amp_n++];
   store->send_input = !store->send_input;
   return ret;
 }
 
-int machine_writer(void* data_ptr, int val) {
+int machine_writer(void* data_ptr, long val) {
   struct Store* store = (struct Store*)data_ptr;
   store->current_val = val;
 
   return 0;
 }
 
-int* ingest(int* size) {
-  int* list = calloc(SIZE, sizeof(int));
+long* ingest(int* size) {
+  long* list = calloc(SIZE, sizeof(long));
   int capacity = SIZE;
   int _size = 0;
   int i;
   while(scanf("%d,?", &i) != EOF) {
     if (_size == capacity) {
       capacity *= 2;
-      list = realloc(list, capacity * sizeof(int));
+      list = realloc(list, capacity * sizeof(long));
     }
